port nonlinear case to new structure
